Recommendations verified by the script
Even if the user provides a recommendation for processing the bib/index
entries, pytex verifies whether it matches with its recommendations. In
case of discrepancy, it warns the user.

In addition, I removed \bibstyle from the regexp used to start
processing bib directives. It might very well happen that someone has a
\bibstyle, but is not producing any citations.