New RE_RERUN regexp
RE_RERUN is applied over the contents of the log file and it helps to
determine whether a new pass of the processor is required. The previous
expression failed to observe that the processor had to be executed again
when there were undefined citations. It is fixed and it now seems to
work as expected in all cases tried so far.